requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Deprecated==1.2.13
Django==4.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
elementpath==3.0.2
exceptiongroup==1.0.4
gunicorn==20.1.0
h11==0.14.0
idna==3.4
kombu==5.2.4
mysqlclient==2.1.1
outcome==1.2.0
packaging==21.3
pi==0.1.2
prompt-toolkit==3.0.31
psutil==5.9.4
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
selenium==4.6.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.6
urllib3==1.26.12
vine==5.0.0
waitress==2.1.2
wcwidth==0.2.5
webdriver-manager==3.8.5
wrapt==1.14.1
wsproto==1.2.0
